Initial set of Xplanet and Totalmarker files
This is technically a branch of Xplanet 1.3.0 with Totalmarker 2.6.1.  I have attempted to create module files of the various subroutines in Totalmarker so that we can maintain (or add) modules for the various earth oriented functions of Xplanet.

CloudUpdate.pm is working with a fetch of cloud maps from Xeric, which requires a subscription.  I moved subroutine 'Cloud_update' into this module and it's the only subroutine for clouds.  I also cleaned up the fetch routine so it is not longer using 'WGET', per Michael's notes in the main script.  I also made the userid and password variables to pass in, so we could manage to control the encryption of the userid and password from being in clear text when commiting.

Hurricanes.pm is not working at all, since the storm data has either moved or been discontinued by the sources previously identified by Michael Dear.  We'll have to find new ones and rebuild this module.  This is driver for moving the subroutines to modules as I was having trouble wrapping my head around a 4,000 line perl script.  Also, Michael Dear wrote that he'd like to have a better cloud map fetching mechanism and I updated that in Clouds.pm.

I had to create a Globals.pm for all the global variables and settings that Michael was using.  I had a hard time tracking all of them down.  I won't list the variables here since they are declared at the head of the module.

Earthquake.pm is the module file for earthquakes.  This is still working, 20 years later.  I moved a lot of subroutines to this module, namely:
- drawcircle
- max_model
- max_min_model
- standar_model
- colourisetext
- colourisemag
- WriteoutQuake
- get_Correct_quake_Feed
- get_quakedata

Volcano.pm contains the following subroutines:
- WriteoutVolcano
- get_volcanodata
- volcanodata_checked

I have not made any changes to the volcano subroutines themselves.  They seem to work fine.

Norad.pm contains three subroutines:
- get_noraddata
- norad_checked
- update_file

I'm not completely clear on why the update_file routine contains settings for the satellites, when one can override them with image settings in the marker file.  Since the shuttle does not fly anymore, some of this is fallow.  I didn't deprecate anything here, and updating the marker files is a lot easier anyways.

Label.pm has the following subroutines:
- WRiteoutlabel
- process_update
- process_last_update
- update_time_status
- num_of_month
- file_header

I did not make any changes (so far) to these routines from Michael's code.

Eclipse.pm contains the following subroutines:
- get_eclipsedata
- readineclipseindex
- readineclipsetrack
- datacurrent
- writeouteclipsemarker
- writeouteclipsearcboarder
- writeouteclipsearccenter
- writeouteclipsefilesnone
- writeouteclipselabel
- refinedata

I have done zero testing of this module.  I don't know if the sources are still working.  I'll have to check with some historical data perhaps to see what happens and if the sources work correctly.

EasterEgg.pm contains the subroutine easteregg.  No testing has been done here either.  It's all about Santa Clause.

I am committing my personal copies of marker files as additional examples of making Xplanet sit up and do tricks.

I previously used to pull fire data from the MODIS site and plot that, so we can monitor wildfires and intentional deforestation activities by humanity.  I intend to add that as a module to Totalmarker.

Eventually, once this is at least working, I intend to commit this as a branch to the original Xplanet codeline.

Lastly, the main script by Michael Dear did not use STRICT or WARNINGS.  Once I started using those, I had a lot of main script updates to do, along with declarations.  I hope that this preserves the original functionality.  May others who come after me build on this, I hope they do.