Earthquake module updates
Had some problems getting the Earthquake module to run.  Added debugging code to the file input from the main script, which had a leading space in it for some reason.  A regex removes that.  Couldn't write the quake marker, which now shows up.

Having some issues with the Hurricane/storm code, which I believe are caused by a prior change by me, namely modifying the code to avoid repeated use of a local $counter, which if I put into Globals becomes a global problem.  Also, I seem to have an extra routine which I don't recall the source.